Add a json-dictionary based map
If there is a place you got to go
I am the one you need to know
I'm the Map!
I'm the Map!
I'm the Map!

If there is a place you got to get
I can get you there I bet
I'm the Map!
I'm the Map!
I'm the Map!

I'm the Map!
I'm the Map!
I'm the Map!

I'm the Map!
I'm the Map!
I'm the Map!

I'm the Map!
I'm the Map!
I'm the Map!