Split the main script into some modules
I'm still learning here... but this is looking to be a little more
manageable.  I might mess with the naming scheme, and organization as I
branch out to more modules, but right now, I can define and share
objects in the 'share.py' and have other classes use globals defined
there.