A quick hack to check a map for loops
If a map contains a loop, then the generation the node was encountered
at will be used and checked as the algorithm floods to all of the nodes.
If a node was previsously visited, but is older than our current
generation, then we'll override it's values with a new set!

TODO: Check all of the output